Refactor List Actions, FormBulkActionButtons, DatagridStyles, and colour LIGHTGREY.
Remove duplicate code.
