+++ packages/employer-client/src/common/components/ListActions/ListActions.tsx
@@ -1,0 +1,13 @@
+import { CreateButton, TopToolbar } from "react-admin"
+
+interface ListActionsProps {
+    createButtonLabel: string
+}
+
+export const ListActions: React.FC<ListActionsProps> = ({ createButtonLabel }) => (
+    <TopToolbar sx={{ paddingTop: "5vh" }}>
+        <div>
+            <CreateButton label={createButtonLabel} />
+        </div>
+    </TopToolbar>
+)
