+++ packages/employer-client/src/common/styles/DatagridStyles.tsx
@@ -1,0 +1,14 @@
+export const DatagridStyles = {
+    "& .column-lock": {
+        padding: "6px 0",
+        "& svg": { verticalAlign: "middle" }
+    },
+    "& .RaDatagrid-rowCell": {
+        textAlign: "left",
+        padding: "0.9em 1.3em"
+    },
+    "& .RaDatagrid-headerCell": {
+        fontWeight: "bold",
+        textAlign: "left"
+    }
+}
